feat: mudamos controllers e inserts
