+++ backend/index.js
@@ -1,7 +1,9 @@
 const express = require('express');
 const cors = require('cors');
 const app = express();
-const controllerL = require('./controllers/controllerLogin (1)')
+const controllerLoginAdm = require('./controllers/controllerLoginAdm')
+const controllerRegister = require('./controllers/controllerRegister')
+const controllerLoginUser = require('./controllers/controllerLoginUser')
 
 const script = require("./models/ademiro/adminInsert")
 // Colocando variaveis de ambiente
@@ -23,19 +25,13 @@ const routes = require('./routes/router');
 app.use('/api', routes)
 // app.use("/produto", produto)
 
-app.post("/logar", (req,res,next) =>{
-    const login = req.body.login
-    const senha = req.body.password
 
-    const tokenGer = gerarToken(login) 
+app.use('/logar', controllerLoginAdm.post)
+app.use('/cadastro', controllerRegister.post)
+app.use('/loginUser', controllerLoginUser.post)
 
-    console.log("Login e senha: " + login + " " + senha + " " +  tokenGer)
 
-    res.json({"token" : tokenGer}).status(200)
 
-})
-
-app.use('/logar', controllerL.post)
 // ouvindo uma porta
 
 const mongoose = require("mongoose");
