+++ backend/models/user/insertUser.js
@@ -3,7 +3,6 @@ const modelUser = require('./userModel');
 async function saveUser(body){
 
     const newUser  = new modelUser({
-        telefone: body.telefone,
         email: body.email,
         senha: body.senha
     });
@@ -16,7 +15,6 @@ async function queryUsuario(query){
 
     //Consulta o usuário no banco
     const userSelected = await modelUser.findOne({
-        telefone: query.telefone,
         email: query.email,
         senha: query.senha
     });
